Fixes #239; 0.2.0 Pocket Core compatibility
- Fixed requesthash mis-match; pocket-core is returning a blank proof objecAqt
in the relay response: https://github.com/pokt-network/pocket-core/issues/742

- Fixed signature mis-match; pocket-core and pocket-js weren't in sync with
the order of variables that went into forming the signature. Re-ordered the
relayProof variables in all points of js code to match the right order.

- Removed the proof object from all relay responses; it isn't needed as we're
the ones that send the proof. We don't need to analyze it on the receipt.

- Added note to fix signature validity check

- Made relay headers an optional boolean switch so they aren't included in
the request hash (as requested by pocket-core team)

- Fixed issue where pocket-js was sending the session block height as the
relayMeta block height. This was causing the error of the request height
being behind the actual chain height when the session was more than 5
blocks old. Added a timer functionality that stores the session creation
timestamp when it is first read from dispatch. That creation timestamp
is used whenever a relay is called to determine the current block height.

- Fixed error message when dispatch URL is invalid